CLN Move to utils
