+++ sklearn/preprocessing/_label.py
@@ -24,6 +24,7 @@ from ..utils.validation import _num_samples
 from ..utils.validation import _deprecate_positional_args
 from ..utils.multiclass import unique_labels
 from ..utils.multiclass import type_of_target
+from ..utils._encode import _encode, _unique
 
 
 __all__ = [
@@ -32,146 +33,6 @@ __all__ = [
     'LabelEncoder',
     'MultiLabelBinarizer',
 ]
-
-
-def _encode(values, *, uniques, check_unknown=True):
-    """Helper function encode values.
-
-    Uses pure python method for object dtype, and numpy method for
-    all other dtypes.
-    The numpy method has the limitation that the `uniques` need to
-    be sorted. Importantly, this is not checked but assumed to already be
-    the case. The calling method needs to ensure this for all non-object
-    values.
-
-    Parameters
-    ----------
-    values : array
-        Values to factorize or encode.
-    uniques : array
-        Uniques are not determined from passed values (this
-        can be because the user specified categories, or because they
-        already have been determined in fit).
-    check_unknown : bool, default True
-        If True, check for values in ``values`` that are not in ``unique``
-        and raise an error. This is ignored for object dtype, and treated as
-        True in this case. This parameter is useful for
-        _BaseEncoder._transform() to avoid calling _encode_check_unknown()
-        twice.
-
-    Returns
-    -------
-    encoded : ndarray
-        Encoded values
-    """
-    if values.dtype == object:
-        table = {val: i for i, val in enumerate(uniques)}
-        try:
-            return np.array([table[v] for v in values])
-        except KeyError as e:
-            raise ValueError(f"y contains previously unseen labels: {str(e)}")
-    else:
-        if check_unknown:
-            diff = _encode_check_unknown(values, uniques)
-            if diff:
-                raise ValueError(f"y contains previously unseen labels: "
-                                 f"{str(diff)}")
-        return np.searchsorted(uniques, values)
-
-
-def _unique_python(values, *, return_inverse):
-    # Only used in _uniques below, see docstring there for details
-    try:
-        uniques = sorted(set(values))
-        uniques = np.array(uniques, dtype=values.dtype)
-    except TypeError:
-        types = sorted(t.__qualname__
-                       for t in set(type(v) for v in values))
-        raise TypeError("Encoders require their input to be uniformly "
-                        f"strings or numbers. Got {types}")
-
-    ret = (uniques, )
-
-    if return_inverse:
-        table = {val: i for i, val in enumerate(uniques)}
-        inverse = np.array([table[v] for v in values])
-        ret += (inverse, )
-
-    if len(ret) == 1:
-        ret = ret[0]
-
-    return ret
-
-
-def _unique(values, *, return_inverse=False):
-    """Helper function to find uniques with support for python objects.
-
-    Uses pure python method for object dtype, and numpy method for
-    all other dtypes.
-
-    Parameters
-    ----------
-    unique : ndarray
-        The sorted uniique values
-
-    unique_inverse : ndarray
-        The indicies to reconstruct the original array from the unique array.
-        Only provided if `return_inverse` is True.
-    """
-    if values.dtype == object:
-        return _unique_python(values, return_inverse=return_inverse)
-    # numerical
-    return np.unique(values, return_inverse=return_inverse)
-
-
-def _encode_check_unknown(values, uniques, return_mask=False):
-    """
-    Helper function to check for unknowns in values to be encoded.
-
-    Uses pure python method for object dtype, and numpy method for
-    all other dtypes.
-
-    Parameters
-    ----------
-    values : array
-        Values to check for unknowns.
-    uniques : array
-        Allowed uniques values.
-    return_mask : bool, default False
-        If True, return a mask of the same shape as `values` indicating
-        the valid values.
-
-    Returns
-    -------
-    diff : list
-        The unique values present in `values` and not in `uniques` (the
-        unknown values).
-    valid_mask : boolean array
-        Additionally returned if ``return_mask=True``.
-
-    """
-    if values.dtype == object:
-        uniques_set = set(uniques)
-        diff = list(set(values) - uniques_set)
-        if return_mask:
-            if diff:
-                valid_mask = np.array([val in uniques_set for val in values])
-            else:
-                valid_mask = np.ones(len(values), dtype=bool)
-            return diff, valid_mask
-        else:
-            return diff
-    else:
-        unique_values = np.unique(values)
-        diff = list(np.setdiff1d(unique_values, uniques, assume_unique=True))
-        if return_mask:
-            if diff:
-                valid_mask = np.in1d(values, uniques)
-            else:
-                valid_mask = np.ones(len(values), dtype=bool)
-            return diff, valid_mask
-        else:
-            return diff
 
 
 class LabelEncoder(TransformerMixin, BaseEstimator):
